move fetchUserStreams to donation services
